feat: ✨ 计划作业执行记录 (#89)
顶部通栏黑夜模式开关
计划作业快捷预览面板

+++ src/backend/NetAdmin.SysComponent.Host/Jobs/ScheduledJob.cs
@@ -87,7 +87,7 @@ public sealed class ScheduledJob : WorkBase<ScheduledJob>, IJob
                                             {
                                                 Duration       = sw.ElapsedMilliseconds
                                               , HttpMethod     = job.HttpMethod
-                                              , HttpStatusCode = rsp.StatusCode
+                                              , HttpStatusCode = (int)rsp.StatusCode
                                               , JobId          = job.Id
                                               , RequestBody    = job.RequestBody
                                               , RequestHeader  = _requestHeader
